Reorganize the files and directories
